+++ src/main/java/application/Figures/Segment.java
@@ -88,7 +88,7 @@ public class Segment extends Figure {
     }
 
     public int getRadius() {
-        return dotP1.getRadius();
+        return (int)dotP1.getRadius();
     }
 
     public void setWidth(double width) {
@@ -198,7 +198,7 @@ public class Segment extends Figure {
         if(((TextField)settings[10]).getText()!="") {
             radius = Integer.parseInt(((TextField)settings[10]).getText()) > 0 ? Integer.parseInt(((TextField)settings[10]).getText()) : 1; 
         } else {
-            radius = dotP1.getRadius();
+            radius = (int)dotP1.getRadius();
         }
         if(((TextField)settings[12]).getText()!="") {
             color = Color.web(((TextField)settings[12]).getText());
